Renamed channel.py to securesocket.py
